requirements_v3_improved.txt
langchain
langchain-core
langchain-community
langchain-chroma
PyMuPDF
Pillow
numpy>=1.26.0
chromadb
google-generativeai
python-dotenv
unstructured
